Fixed broken level 200 congratulation (reported and tested by lordstyx)
git-svn-id: https://dev.chiasoft.net:8443/svn/svn_vana/trunk@2469 23f3cdc0-e6d1-42be-a0c1-51f4125517bb

+++ src/ChannelServer/Levels.cpp
@@ -130,7 +130,7 @@ void Levels::giveExp(Player *player, uint32_t exp, bool inChat, bool white) {
 				message = "[Congrats] ";
 				message += player->getName();
 				message += " has reached Level ";
-				message += boost::lexical_cast<string>(getMaxLevel(fulljob));
+				message += boost::lexical_cast<string>((int16_t) getMaxLevel(fulljob));
 				message += "! Congratulate ";
 				message += player->getName();
 				message += " on such an amazing achievement!";
